Move `NeptuneLegacyLogger` to separate directory.
